Do not attempt to reuse last_found_rows_result
A second FOUND_ROWS() query results in a fatal error because
the mysqli_result object stored in $this->result and
$this->last_found_rows_result is the same object after a successful
first call to FOUND_ROWS(), which on the next query is freed
by $this->flush().

This change ensures that a $last_found_rows_result is used only
once and we do not attempt to use a mysqli_result that has been
freed.